changed float to decimal

+++ PES.Data/Day.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,9 +10,16 @@ namespace PES.Data
     
     public class Day
     {
+        //[Key]
         public int DayId { get; set; }
+
+        //[Required]
         public Guid ManagerId { get; set; }
+
+        //[Required]
         public DayOfWeek DayOfWeek { get; set; }
-        public float Sales { get; set; }
+
+        //[Required]
+        public decimal Sales { get; set; }
     }
 }
